Working on refactoring RequestForm component.
